Hotfix 3

+++ install.sh
@@ -53,32 +53,33 @@ sudo mkdir /mnt/drives /mnt/merged
 echo "[*] Please ensure to have done your partitionning before the script execution. CTRL+C if that has not be done yet"
 echo "[#] Here are all partitions :"
 lsblk -o NAME,SIZE
-part="1"
+inputed_part="1"
 id="1"
-while [[ "$part" ]];do
+while [[ "$inputed_part" ]];do
     part=""
-    while [[ ! $part ]];do
+    while [[ $inputed_part && ! $part ]];do
         echo "[*] Select partition :"
         read -p "[>] " inputed_part
-        part="$(ls dev | grep ^$inputed_part$)"
-        if [[ ! $part ]];then
+        part="$(ls /dev | grep ^$inputed_part$)"
+        if [[ $inputed_part && ! $part ]];then
             echo "[!] Invalid partition"
         fi
     done
+    if [[ $inputed_part ]];then
+        type=""
+        while [[ ! $type ]];do
+            echo "[*] Type (vram, hot, cold, temp_hot, temp_cold):"
+            read -p "[>] " inputed_type
+            type=$(echo -e "vram\nhot\ncold\ntemp_hot\ntemp_cold\nthot\ntcold" | grep ^$inputed_type)
+            if [[ ! $type || $(echo $type) != $(echo "$type") ]];then
+                echo "[!] Invalid type"
+            fi
+        done
 
-    type=""
-    while [[ ! $type ]];do
-        echo "[*] Type (vram, hot, cold, temp_hot, temp_cold):"
-        read -p "[>] " inputed_type
-        type=$(echo -e "vram\nhot\ncold\ntemp_hot\ntemp_cold\nthot\ntcold" | grep ^$inputed_type)
-        if [[ ! $type || $(echo $type) != $(echo "$type") ]];then
-            echo "[!] Invalid type"
-        fi
-    done
-
-    sudo mkdir /mnt/drives/$type$id
-    echo "/dev/$part /mnt/drives/$type$id ext4 defaults 0 2" | sudo tee -a /etc/fstab > /dev/null
-    id=$((id+1))
+        sudo mkdir /mnt/drives/$type$id
+        echo "/dev/$part /mnt/drives/$type$id ext4 defaults 0 2" | sudo tee -a /etc/fstab > /dev/null
+        id=$((id+1))
+    fi
 done
 # configure mergerfs
 ## /mnt/merged/vram is used for vram, /mnt/storage is under RAID, /mnt/temp is not, hot is for caching (SSD), cold for archivage (HDD)
@@ -92,14 +93,14 @@ echo "/mnt/drives/temp_cold*:/mnt/drives/tcold* /mnt/merged/temp_cold fuse.merge
 echo "/mnt/drives/temp_hot*:/mnt/drives/thot* /mnt/merged/temp_hot fuse.mergerfs defaults,allow_other,use_ino,cache.files=off,moveonenospc=true,category.create=mfs 0 0" | sudo tee -a /etc/fstab > /dev/null
 echo "/mnt/merged/temp_hot:/mnt/merged/temp_cold /mnt/temp fuse.mergerfs defaults,allow_other,use_ino,cache.files=off,moveonenospc=true,category.create=mfs 0 0" | sudo tee -a /etc/fstab > /dev/null
 # mount
-sudo mount -a
+sudo mount -a 2>/dev/null
 # configure swap
 echo "[~] Configuring swap"
 size=$(df -h /mnt/merged/vram | tail -n1 | awk '{print($4)}')
-sudo fallocate -l $size /mnt/merged/vram/swapfile
-sudo chmod 600 /mnt/merged/vram/swapfile
-sudo mkswap /mnt/merged/vram/swapfile
-sudo swapon /mnt/merged/vram/swapfile
+sudo fallocate -l $size /mnt/merged/vram/swapfile 2>/dev/null
+sudo chmod 600 /mnt/merged/vram/swapfile 2>/dev/null
+sudo mkswap /mnt/merged/vram/swapfile 2>/dev/null
+sudo swapon /mnt/merged/vram/swapfile 2>/dev/null
 echo "/mnt/merged/vram/swapfile none swap sw 0 0" | sudo tee -a /etc/fstab > /dev/null
 # hot-cold storage management - TODO
 
@@ -114,37 +115,37 @@ for py in $(ls /usr/lib/ | grep python3.);do
         sudo mv /usr/lib/$py/EXTERNALLY-MANAGED /usr/lib/$py/EXTERNALLY-MANAGED.old
     fi
 done
-pip3 install frida
+pip3 install frida -q >/dev/null 2>/dev/null
 
 echo "[~] Fetching script"
-git clone https://github.com/DualCoder/vgpu_unlock
+git clone https://github.com/DualCoder/vgpu_unlock --quiet >/dev/null 2>/dev/null
 chmod -R +x vgpu_unlock
 sudo mv vgpu_unlock /lib/
 
 echo "[~] Setting up iommu"
 sudo sed -i 's/GRUB_CMDLINE_LINUX_DEFAULT="quiet"/GRUB_CMDLINE_LINUX_DEFAULT="quiet intel_iommu=on iommu=pt"/' /etc/default/grub
-sudo update-grub
+sudo update-grub >/dev/null 2>/dev/null
 
 echo -e "\nvfio\nvfio_iommu_typel\nvfio_pci\nvfio_virqfd\n" | sudo tee -a /etc/modules >/dev/null
 echo "options vfio_iommu_typel allow_unsafe_interrupts=1" | sudo tee /etc/modprobe.d/iommu_unsafe_interrupts.conf >/dev/null
 echo "options kvm ignore_msrs=1" | sudo tee /etc/modprobe.d/kvm_msrs.conf >/dev/null
 echo "blacklist nouveau" | sudo tee -a /etc/modprobe.d/blacklist.conf >/dev/null
-sudo update-initramfs -u
+sudo update-initramfs -u >/dev/null 2>/dev/null
 
 echo "[~] Fetching Drivers"
 # https://github.com/wvthoog/proxmox-vgpu-installer/blob/main/proxmox-installer.sh
 version="550.54.10"
-megadl https://mega.nz/file/JjtyXRiC#cTIIvOIxu8vf-RdhaJMGZAwSgYmqcVEKNNnRRJTwDFI
-chmod +x NVIDIA-Linux-x86_64-$version-vgpu-kvm
-sudo ./NVIDIA-Linux-x86_64-$version-vgpu-kvm --dkms
+megadl https://mega.nz/file/JjtyXRiC#cTIIvOIxu8vf-RdhaJMGZAwSgYmqcVEKNNnRRJTwDFI >/dev/null 2>/dev/null
+chmod +x NVIDIA-Linux-x86_64-$version-vgpu-kvm.run
+sudo ./NVIDIA-Linux-x86_64-$version-vgpu-kvm.run --dkms
 sudo sed -i 's/ExecStart=/ExecStart=\/lib\/vgpu_unlock\/vgpu_unlock /' /lib/systemd/system/nvidia-vgpud.service
 sudo sed -i 's/ExecStart=/ExecStart=\/lib\/vgpu_unlock\/vgpu_unlock /' /lib/systemd/system/nvidia-vgpu-mgr.service
 systemctl daemon-reload
 sudo sed -i 's/cpuset.h>/cpuset.h>\n#include "\/lib\/vgpu_unlock\/vgpu_unlock_hooks.c"/' /usr/src/nvidia-$version/nvidia/os-interface.c
 echo "ldflags-y += -T /lib/vgpu_unlock/kern.ld" | sudo tee -a /usr/src/nvidia-$version/nvidia/nvidia.Kbuild >/dev/ull
 echo "[~] Building driver"
-dkms remove -m nvidia -v $version --all
-dkms install -m nvidia -v $version
+dkms remove -m nvidia -v $version --all >/dev/null 2>/dev/null
+dkms install -m nvidia -v $version >/dev/null 2>/dev/null
 
 # Proxmox installation
 ## Hostname management
